move tests to correct file
